+++ components/forms/farmer-journal/ChickenStock.tsx
@@ -1,5 +1,5 @@
 import { useFormContext } from "react-hook-form";
-import ErrorMessage from "../ErrorMeesge";
+import ErrorMessage from "../ErrorMeesage";
 import Label from "../Label";
 import PreviewField from "../PreviewField";
 import NumberInput from "./NumberInput";
